% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_detector.R
\name{run_detector}
\alias{run_detector}
\title{Run Detector}
\usage{
run_detector(0.3, "cfg/yolo.weights", "cfg/coco.dat", "cfg/yolo.cfg",
                   "data/dog.jpg", "out.jpg")
}
\arguments{
\item{thresh}{Threshold to use}

\item{weight_file}{File with the weights}

\item{dat_file}{Dat file}

\item{conf_file}{Configuration file}

\item{input_file}{Input image file}

\item{output_file}{Output image file}
}
\description{
Runs object detection on image file using a pre-trained model.
}
